% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroepirecipes-package.R
\docType{package}
\name{seroepirecipes-package}
\alias{seroepirecipes}
\alias{seroepirecipes-package}
\title{seroepirecipes: seroepirecipes: a codebase for seroanalytics methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}


}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seroanalytics/seroepirecipes}
  \item Report bugs at \url{https://github.com/seroanalytics/seroepirecipes/issues}
}

}
\keyword{Rcpp}
\keyword{extract}
\keyword{import}
\keyword{importFrom}
\keyword{internal}
\keyword{methods}
\keyword{rstan}
\keyword{sampling}
