% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroepirecipes-package.R
\docType{package}
\name{seroepirecipes-package}
\alias{seroepirecipes-package}
\alias{_PACKAGE}
\alias{seroepirecipes}
\title{seroepirecipes}
\description{
seroepirecipes: implementations of seroanalytical methods
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seroanalytics/seroepirecipes}
  \item Report bugs at \url{https://github.com/seroanalytics/seroepirecipes/issues}
}

}
